+++ docs/pages/release_notes.md
@@ -25,6 +25,9 @@ This is a {{ site.pmd.release_type }} release.
 ### 🚀️ New and noteworthy
 
 ### 🐛️ Fixed Issues
+* java-codestyle
+    * [#5732](https://github.com/pmd/pmd/issues/5732): \[java] UnnecessaryCast: False positive for package-private members on subclasses in another package
+
 
 ### 🚨️ API Changes